Refactored online.py into tutorlib/online
Only very minor changes.  Broke up a few parts just to be consistent
with how the rest of MPT is structured.

Minor changes to SessionManager API (no longer need to leak the SEVER
url from the file).
